gdbus-codegen: Make header inclusion simply relative
Otherwise in "complex" cases, for example using glib as a subproject
in meson the output dir is relative to the toplevel build directory
(meaning subprojects/glib/gio/...) and the include paths won't fit.

The header file and C file being always generated in the same output
directory, the simplest solution is to just do a relative import using
the header filename.